pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "discord-py>=2.5.2",
    "pillow>=11.1.0",
    "pynacl>=1.5.0",
    "pyopenssl>=25.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "youtube-dl>=2021.12.17",
    "yt-dlp>=2025.2.19",
]
